+++ lib/packages/contracts/src/constants.ts
@@ -1,0 +1,10 @@
+/** @module contracts */
+
+/**
+ * Copyright (c) 2019 Burst Apps Team
+ */
+
+/**
+ * Size of Code Pages in bytes
+ */
+export const CodePageSize = 256;