+++ lib/packages/core/src/api/factories/contract/index.ts
@@ -3,3 +3,4 @@
 export * from './getContract';
 export * from './getAllContractIds';
 export * from './getContractsByAccount';
+export * from './publishContract';