+++ lib/packages/core/src/api/factories/network/suggestFee.ts
@@ -5,7 +5,7 @@
  */
 import {BurstService} from '../../../service/burstService';
 import {SuggestedFees} from '../../../typings/suggestedFees';
-import {FeeQuantNQT} from '../../../constants';
+import {FeeQuantPlanck} from '../../../constants';
 
 /**
  * Use with [[ApiComposer]] and belongs to [[NetworkApi]].
@@ -17,7 +17,7 @@ export const suggestFee = (service: BurstService): () => Promise<SuggestedFees> 
         const suggestedFees: SuggestedFees = await service.query('suggestFee');
         return {
             ...suggestedFees,
-            minimum: FeeQuantNQT
+            minimum: FeeQuantPlanck
         };
     };
 };