+++ lib/packages/core/src/contract/index.ts
@@ -1,4 +1,0 @@
-/** @module core */
-
-export * from './getContractDatablock';
-export * from './ContractHelper';