+++ lib/packages/core/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.5.0
+
+___
+__Breaking Changes__
+- `ContractHelper`, `getContractDatablock` moved into new package `contracts`
+- `FeeQuantNQT` renamed to `FeeQuantPlanck`
+___
+
+- added new methods to contract Api
+    - `callContractMethod`
+    - `publishContract`
+
 ## 0.4.3
 ___
 __Breaking Changes__